chore: moves TransportError to separate file
