+++ ts/src/sdk/chain/server/index.ts
@@ -1,4 +1,4 @@
 export * from "./createChainNodeSDK.ts";
 export type { TxClient } from "../../transport/tx/TxClient.ts";
-export { TransportError as ChainSDKError } from "../../transport/transportUtils.ts";
+export { TransportError as SDKError } from "../../transport/TransportError.ts";
 export { TxError } from "../../transport/tx/TxError.ts";
