+++ ts/src/sdk/provider/server/index.ts
@@ -1,4 +1,4 @@
 export * from "./createProviderSDK.ts";
 export * from "../auth/mtls/index.ts";
 export * from "../auth/jwt/index.ts";
-export { TransportError as ChainSDKError } from "../../transport/transportUtils.ts";
+export { TransportError as SDKError } from "../../transport/TransportError.ts";
